+++ cryptofeed/exchange/okcoin.py
@@ -4,19 +4,24 @@ Copyright (C) 2017-2021  Bryant Moscon - bmoscon@gmail.com
 Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
+import asyncio
+
 from decimal import Decimal
 from itertools import islice
+from functools import partial
 import logging
 import zlib
+from typing import List, Tuple, Callable
 
 from sortedcontainers import SortedDict as sd
 from yapic import json
 
+from cryptofeed.auth.okcoin import generate_token
 from cryptofeed.connection import AsyncConnection
-from cryptofeed.defines import ASK, BID, BUY, FUNDING, L2_BOOK, OKCOIN, OPEN_INTEREST, SELL, TICKER, TRADES, LIQUIDATIONS
+from cryptofeed.defines import ASK, BID, BUY, FUNDING, L2_BOOK, OKCOIN, OPEN_INTEREST, SELL, TICKER, TRADES, LIQUIDATIONS, ORDER_INFO
 from cryptofeed.exceptions import BadChecksum
 from cryptofeed.feed import Feed
-from cryptofeed.standards import symbol_exchange_to_std, timestamp_normalize
+from cryptofeed.standards import symbol_exchange_to_std, timestamp_normalize, is_authenticated_channel
 from cryptofeed.symbols import get_symbol_separator
 from cryptofeed.util import split
 
@@ -85,13 +90,14 @@ class OKCoin(Feed):
 
     def get_channel_symbol_combinations(self):
         for chan in set(self.channels or self.subscription):
-            if chan == LIQUIDATIONS:
-                continue
-            for symbol in set(self.symbols or self.subscription[chan]):
-                instrument_type = self.instrument_type(symbol)
-                if instrument_type != 'swap' and 'funding' in chan:
-                    continue  # No funding for spot, futures and options
-                yield f"{chan.format(instrument_type)}:{symbol}"
+            if not is_authenticated_channel(chan):
+                if chan == LIQUIDATIONS:
+                    continue
+                for symbol in set(self.symbols or self.subscription[chan]):
+                    instrument_type = self.instrument_type(symbol)
+                    if instrument_type != 'swap' and 'funding' in chan:
+                        continue  # No funding for spot, futures and options
+                    yield f"{chan.format(instrument_type)}:{symbol}"
 
     async def _ticker(self, msg: dict, timestamp: float):
         """
@@ -102,8 +108,8 @@ class OKCoin(Feed):
             update_timestamp = timestamp_normalize(self.id, update['timestamp'])
             await self.callback(TICKER, feed=self.id,
                                 symbol=pair,
-                                bid=Decimal(update['best_bid']),
-                                ask=Decimal(update['best_ask']),
+                                bid=Decimal(update['best_bid']) if update['best_bid'] else Decimal(0),
+                                ask=Decimal(update['best_ask']) if update['best_ask'] else Decimal(0),
                                 timestamp=update_timestamp,
                                 receipt_timestamp=timestamp)
             if 'open_interest' in update:
@@ -118,7 +124,7 @@ class OKCoin(Feed):
         {'table': 'spot/trade', 'data': [{'instrument_id': 'BTC-USD', 'price': '3977.44', 'side': 'buy', 'size': '0.0096', 'timestamp': '2019-03-22T22:45:44.578Z', 'trade_id': '486519521'}]}
         """
         for trade in msg['data']:
-            if msg['table'] == 'futures/trade':
+            if msg['table'] in {'futures/trade', 'option/trade'}:
                 amount_sym = 'qty'
             else:
                 amount_sym = 'size'
@@ -182,6 +188,29 @@ class OKCoin(Feed):
                     raise BadChecksum
                 await self.book_callback(self.l2_book[pair], L2_BOOK, pair, False, delta, timestamp_normalize(self.id, update['timestamp']), timestamp)
 
+    async def _order(self, msg: dict, timestamp: float):
+        if msg['data'][0]['status'] == "open":
+            status = "active"
+        else:
+            status = msg['data'][0]['status']
+
+        keys = ('filled_size', 'size', 'filled_notional')
+        data = {k: Decimal(msg['data'][0][k]) for k in keys if k in msg['data'][0]}
+
+        await self.callback(ORDER_INFO, feed=self.id,
+                            symbol=symbol_exchange_to_std(msg['data'][0]['instrument_id'].upper()),  # This uses the REST endpoint format (lower case)
+                            status=status,
+                            order_id=msg['data'][0]['order_id'],
+                            side=BUY if msg['data'][0]['side'].lower() == 'buy' else SELL,
+                            order_type=msg['data'][0]['type'],
+                            timestamp=msg['data'][0]['timestamp'].timestamp(),
+                            receipt_timestamp=timestamp,
+                            **data
+                            )
+
+    async def _login(self, msg: dict, timestamp: float):
+        LOG.info('%s: Websocket logged in? %s', self.id, msg['success'])
+
     async def message_handler(self, msg: str, conn, timestamp: float):
 
         # DEFLATE compression, no header
@@ -193,6 +222,8 @@ class OKCoin(Feed):
                 LOG.error("%s: Error: %s", self.id, msg)
             elif msg['event'] == 'subscribe':
                 pass
+            elif msg['event'] == 'login':
+                await self._login(msg, timestamp)
             else:
                 LOG.warning("%s: Unhandled event %s", self.id, msg)
         elif 'table' in msg:
@@ -204,7 +235,32 @@ class OKCoin(Feed):
                 await self._book(msg, timestamp)
             elif 'swap/funding_rate' in msg['table']:
                 await self._funding(msg, timestamp)
+            elif 'spot/order' in msg['table']:
+                await self._order(msg, timestamp)
             else:
                 LOG.warning("%s: Unhandled message %s", self.id, msg)
         else:
             LOG.warning("%s: Unhandled message %s", self.id, msg)
+
+    def connect(self) -> List[Tuple[AsyncConnection, Callable[[None], None], Callable[[str, float], None]]]:
+        ret = []
+        for channel in self.subscription or self.channels:
+            if is_authenticated_channel(channel):
+                syms = self.symbols or self.subscription[channel]
+                for s in syms:
+                    ret.append((AsyncConnection(self.address, self.id, **self.ws_defaults), partial(self.user_order_subscribe, symbol=s), self.message_handler))
+            else:
+                ret.append((AsyncConnection(self.address, self.id, **self.ws_defaults), self.subscribe, self.message_handler))
+
+        return ret
+
+    async def user_order_subscribe(self, conn: AsyncConnection, symbol=None):
+        self.__reset()
+        timestamp, sign = generate_token(self.key_id, self.key_secret)
+        login_param = {"op": "login", "args": [self.key_id, self.config.okex.key_passphrase, timestamp, sign.decode("utf-8")]}
+        login_str = json.dumps(login_param)
+        await conn.send(login_str)
+        await asyncio.sleep(5)
+        sub_param = {"op": "subscribe", "args": ["spot/order:{}".format(symbol)]}
+        sub_str = json.dumps(sub_param)
+        await conn.send(sub_str)