+++ examples/demo_oi_liq.py
@@ -55,7 +55,7 @@ def main():
 
     f.add_feed(BinanceFutures(symbols=binance_futures_symbols(), channels=[OPEN_INTEREST, LIQUIDATIONS], callbacks={OPEN_INTEREST: OpenInterestCallback(oi), LIQUIDATIONS: LiquidationCallback(liquidations)}))
 
-    f.add_feed(Deribit(symbols=['BTC-PERPETUAL', 'ETH-PERPETUAL'], channels=[LIQUIDATIONS, OPEN_INTEREST],
+    f.add_feed(Deribit(symbols=['BTC-USD-PERPETUAL', 'ETH-USD-PERPETUAL'], channels=[LIQUIDATIONS, OPEN_INTEREST],
                        callbacks={OPEN_INTEREST: OpenInterestCallback(oi),
                                   LIQUIDATIONS: LiquidationCallback(liquidations)}))
     f.run()